move existing components to "components" folder
