+++ src/routers/Search/dashboard/src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as  ArtifactSetViewer } from "./ArtifactSetViewer.svelte";
+export { default as TaxonomyFilter } from "./TaxonomyFilter.svelte";
+export { default as TaxonomyFilterTree } from "./TaxonomyFilterTree.svelte";
